requirements.txt
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
docopt==0.6.2
fonttools==4.46.0
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
pdfminer.six==20220524
Pillow==10.1.0
py-pdf-parser==0.12.0
pycparser==2.21
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
pyvoronoi==1.0.8.4
shapely==2.0.2
six==1.16.0
Wand==0.6.10